+++ glean-core/ios/GleanTests/UploaderCapabilityTests.swift
@@ -6,22 +6,21 @@
 import XCTest
 
 // REASON: This test is long because of setup boilerplate
-class HttpPingUploaderTests: XCTestCase {
+class UploaderCapabilityTests: XCTestCase {
     var expectation: XCTestExpectation?
-    private let testPath = "/some/random/path/not/important"
-    private let testPing = "{ \"ping\": \"test\" }"
-    private let testRequest = PingRequest(
+
+    private let testOhttpRequest = PingRequest(
         documentId: "12345",
         path: "/some/random/path/not/important",
         body: [UInt8]("{ \"ping\": \"test\" }".utf8),
         headers: [:],
         bodyHasInfoSections: true,
         pingName: "testPing",
-        uploaderCapabilities: []
+        uploaderCapabilities: ["ohttp"]
     )
 
     override func setUp() {
-        resetGleanDiscardingInitialPings(testCase: self, tag: "HttpPingUploaderTests", clearStores: true)
+        resetGleanDiscardingInitialPings(testCase: self, tag: "UploaderCapabilityTests", clearStores: true)
     }
 
     override func tearDown() {
@@ -33,16 +32,27 @@ class HttpPingUploaderTests: XCTestCase {
     /// Launch a new ping uploader on the background thread.
     ///
     /// This function doesn't block.
-    private func getUploader() -> HttpPingUploader {
+    private func getUploader(_ capabilities: [String] = []) -> HttpPingUploader {
         // Build a URLSession with no-caching suitable for uploading our pings
         let config: URLSessionConfiguration = .default
         config.requestCachePolicy = NSURLRequest.CachePolicy.reloadIgnoringLocalCacheData
         config.urlCache = nil
         let session = URLSession(configuration: config)
-        return HttpPingUploader.init(configuration: Configuration(), session: session)
+        return HttpPingUploader.init(configuration: Configuration(), session: session, capabilities: capabilities)
     }
 
-    func testHTTPStatusCode() {
+    func testUploaderIncapable() {
+        let httpPingUploader = self.getUploader()
+        httpPingUploader.upload(request: testOhttpRequest) { result in
+            XCTAssertEqual(
+                .incapable(unused: 0),
+                result,
+                "upload should return .incapable when capabilities don't match"
+            )
+        }
+    }
+
+    func testUploaderCapable() {
         // We are explicitly setting the test mode to true here to force the uploader to not
         // run in the background, which can make this test take a long time.
         var testValue: UploadResult?
@@ -50,13 +60,10 @@ class HttpPingUploaderTests: XCTestCase {
             XCTAssert(json != nil)
             XCTAssertEqual(json?["ping"] as? String, "test")
         }
-
         expectation = expectation(description: "Completed upload")
 
-        // Build a URLSession with no-caching suitable for uploading our pings
-
-        let httpPingUploader = self.getUploader()
-        httpPingUploader.upload(request: testRequest) { result in
+        let ohttpCapableUploader = self.getUploader(["ohttp"])
+        ohttpCapableUploader.upload(request: testOhttpRequest) { result in
             testValue = result
             self.expectation?.fulfill()
         }
@@ -72,45 +79,5 @@ class HttpPingUploaderTests: XCTestCase {
             let value = String(describing: testValue)
             XCTAssertTrue(false, "`upload()` returns the expected HTTP status code. Was: \(value)")
         }
-    }
-
-    func testRequestParameters() {
-        // Build a URLSession with no-caching suitable for uploading our pings
-        let config: URLSessionConfiguration = .default
-        config.requestCachePolicy = NSURLRequest.CachePolicy.reloadIgnoringLocalCacheData
-        config.urlCache = nil
-        let session = URLSession(configuration: config)
-
-        // Build a request.
-        // We specify a single additional header here.
-        // In usual code they are generated on the Rust side.
-        let request = HttpPingUploader(configuration: Configuration(), session: session)
-            .buildRequest(path: testPath, data: Data(testPing.utf8), headers: ["X-Test-Only": "Glean"])
-
-        XCTAssertEqual(
-            request?.url?.path,
-            testPath,
-            "Request path set correctly"
-        )
-        XCTAssertEqual(
-            request?.httpMethod,
-            "POST",
-            "Request method set correctly"
-        )
-        XCTAssertEqual(
-            request?.httpBody,
-            Data(testPing.utf8),
-            "Request body set correctly"
-        )
-        XCTAssertEqual(
-            request?.httpShouldHandleCookies,
-            false,
-            "Request cookie policy set correctly"
-        )
-        XCTAssertEqual(
-            request?.value(forHTTPHeaderField: "X-Test-Only"),
-            "Glean",
-            "Request header set correctly"
-        )
     }
 }