+++ glean-core/rlb/examples/crashing-threads.rs
@@ -98,6 +98,7 @@ pub static PrototypePing: Lazy<PingType> = Lazy::new(|| {
         vec![],
         vec![],
         true,
+        vec![],
     )
 });